refactor(command-palette): absorb shortcut and reset into shared component, eliminate wrappers
CommandPalette now owns Cmd/Ctrl+K toggling (configurable via shortcut
prop, null to disable) and auto-resets search on close. This removes
the duplicated svelte:window handlers and reset effects from every
consumer.

Delete wrapper components that just re-exported CommandPalette under the
same name—each app now imports directly from @epicenter/ui/command-palette
and passes items inline. Skills sidebar header replaces the Toolbar with
an integrated search trigger.
