+++ apps/opensidian/src/lib/components/AppShell.svelte
@@ -1,12 +1,44 @@
 <script lang="ts">
 	import * as Resizable from '@epicenter/ui/resizable';
 	import { ScrollArea } from '@epicenter/ui/scroll-area';
+	import { CommandPalette, type CommandPaletteItem } from '@epicenter/ui/command-palette';
+	import type { FileId } from '@epicenter/filesystem';
 	import { terminalState } from '$lib/state/terminal-state.svelte';
-	import CommandPalette from './CommandPalette.svelte';
+	import { fsState } from '$lib/state/fs-state.svelte';
+	import { getFileIcon } from '$lib/utils/file-icons';
 	import ContentPanel from './editor/ContentPanel.svelte';
 	import Toolbar from './Toolbar.svelte';
 	import TerminalPanel from './terminal/TerminalPanel.svelte';
 	import FileTree from './tree/FileTree.svelte';
+
+	let paletteOpen = $state(false);
+
+	// \u2500\u2500 Collect all files recursively (only when palette is open) \u2500\u2500\u2500\u2500
+	type FileEntry = { id: FileId; name: string; parentDir: string };
+
+	const allFiles = $derived.by((): FileEntry[] => {
+		if (!paletteOpen) return [];
+		return fsState.walkTree<FileEntry>((id, row) => {
+			if (row.type === 'file') {
+				const fullPath = fsState.getPathForId(id) ?? '';
+				const lastSlash = fullPath.lastIndexOf('/');
+				const parentDir = lastSlash > 0 ? fullPath.slice(1, lastSlash) : '';
+				return { collect: { id, name: row.name, parentDir }, descend: false };
+			}
+			return { descend: true };
+		});
+	});
+
+	const fileItems = $derived<CommandPaletteItem[]>(
+		allFiles.map((file) => ({
+			id: file.id,
+			label: file.name,
+			description: file.parentDir || undefined,
+			icon: getFileIcon(file.name),
+			group: 'Files',
+			onSelect: () => fsState.selectFile(file.id),
+		})),
+	);
 
 	let terminalRef: ReturnType<typeof TerminalPanel> | undefined = $state();
 	let previousFocus: HTMLElement | null = $state(null);
@@ -65,6 +97,13 @@
 			</Resizable.PaneGroup>
 		</Resizable.Pane>
 	</Resizable.PaneGroup>
-	<CommandPalette />
+	<CommandPalette
+		items={fileItems}
+		bind:open={paletteOpen}
+		placeholder="Search files..."
+		emptyMessage="No files found."
+		title="Search Files"
+		description="Search for a file by name"
+	/>
 </div>
