+++ apps/tab-manager/src/entrypoints/sidepanel/App.svelte
@@ -12,7 +12,8 @@
 	import ZapIcon from '@lucide/svelte/icons/zap';
 	import { onMount } from 'svelte';
 	import AiDrawer from '$lib/components/AiDrawer.svelte';
-	import { CommandPalette } from '$lib/components/command-palette';
+	import { CommandPalette } from '@epicenter/ui/command-palette';
+	import { items as commandPaletteItems } from '$lib/components/command-palette/items';
 	import SyncStatusIndicator from '$lib/components/SyncStatusIndicator.svelte';
 	import UnifiedTabList from '$lib/components/tabs/UnifiedTabList.svelte';
 	import { browserState } from '$lib/state/browser-state.svelte';
@@ -127,5 +128,5 @@
 	</main>
 </Tooltip.Provider>
 <ConfirmationDialog />
-<CommandPalette bind:open={commandPaletteOpen} />
+	<CommandPalette items={commandPaletteItems} bind:open={commandPaletteOpen} shortcut={null} />
 <AiDrawer bind:open={aiDrawerOpen} />
