--- apps/tab-manager/src/lib/components/command-palette/CommandPalette.svelte
@@ -1,8 +1,0 @@
-<script lang="ts">
-	import { CommandPalette } from '@epicenter/ui/command-palette';
-	import { items } from './items';
-
-	let { open = $bindable(false) }: { open: boolean } = $props();
-</script>
-
-<CommandPalette {items} bind:open />
