--- apps/tab-manager/src/lib/components/command-palette/index.ts
@@ -1,1 +1,0 @@
-export { default as CommandPalette } from './CommandPalette.svelte';
